feat(redis): add downtime threshold before reporting nodes as offline
BREAKING CHANGE: Nodes are no longer just stored as `offline` = `0` or `1`. This
is now an incrementing field for the number of times a node has been seen
offline in a row. Nodes are only reported offline once they have been offline
for `OFFLINE_THRES` times and will not be reported as online unless this
threashold was reached.